seperate seeding from index
